+++ lib/src/manager/im_conversation_manager.dart
@@ -204,6 +204,34 @@ class ConversationManager {
               }))
           .then((value) => Utils.toListMap(value));
 
+  /// burn after reading
+  /// 阅后即焚
+  Future<dynamic> setOneConversationPrivateChat({
+    required String conversationID,
+    required bool isPrivate,
+    String? operationID,
+  }) =>
+      _channel.invokeMethod(
+          'setOneConversationPrivateChat',
+          _buildParam({
+            "conversationID": conversationID,
+            "isPrivate": isPrivate,
+            "operationID": Utils.checkOperationID(operationID),
+          }));
+
+  /// Delete conversation from local and service
+  /// 删除会话
+  Future<dynamic> deleteConversationMsgFromLocalAndSvr({
+    required String conversationID,
+    String? operationID,
+  }) =>
+      _channel.invokeMethod(
+          'deleteConversationMsgFromLocalAndSvr',
+          _buildParam({
+            "conversationID": conversationID,
+            "operationID": Utils.checkOperationID(operationID),
+          }));
+
   /// Custom sort for conversation list
   /// 会话列表自定义排序规则。
   List<ConversationInfo> simpleSort(List<ConversationInfo> list) => list