+++ lib/src/manager/im_message_manager.dart
@@ -56,10 +56,12 @@ class MessageManager {
   /// Find all history message
   /// 获取聊天记录
   /// [userID]接收消息的用户id
+  /// [conversationID] 会话id，查询通知是可用
   /// [groupID]接收消息的组id
   Future<List<Message>> getHistoryMessageList({
     String? userID,
     String? groupID,
+    String? conversationID,
     Message? startMsg,
     int? count,
     String? operationID,
@@ -70,6 +72,7 @@ class MessageManager {
               _buildParam({
                 'userID': userID ?? '',
                 'groupID': groupID ?? '',
+                'conversationID': conversationID ?? '',
                 'startClientMsgID': startMsg?.clientMsgID ?? '',
                 'count': count ?? 10,
                 'operationID': Utils.checkOperationID(operationID),
@@ -522,35 +525,98 @@ class MessageManager {
 
   /// Search local message
   /// 搜索消息
-  Future<dynamic> searchLocalMessages({
+  /// [sourceID]单聊为用户ID，群聊为群ID
+  /// [sessionType]会话类型，单聊为1，群聊为2，如果为0，则代表搜索全部
+  /// [keywordList]搜索关键词列表，目前仅支持一个关键词搜索
+  /// [keywordListMatchType]关键词匹配模式，1代表与，2代表或，暂时未用
+  /// [senderUserIDList]指定消息发送的uid列表 暂时未用
+  /// [messageTypeList]消息类型列表
+  /// [searchTimePosition]搜索的起始时间点。默认为0即代表从现在开始搜索。UTC 时间戳，单位：秒
+  /// [searchTimePeriod]从起始时间点开始的过去时间范围，单位秒。默认为0即代表不限制时间范围，传24x60x60代表过去一天
+  /// [pageIndex]当前页数
+  /// [count]每页数量
+  Future<SearchResult> searchLocalMessages({
     required String sourceID,
     required int sessionType,
     List<String> keywordList = const [],
     int keywordListMatchType = 0,
     List<String> senderUserIDList = const [],
-    List<String> messageTypeList = const [],
+    List<int> messageTypeList = const [],
     int searchTimePosition = 0,
     int searchTimePeriod = 0,
     int pageIndex = 1,
     int count = 40,
     String? operationID,
   }) =>
+      _channel
+          .invokeMethod(
+              'searchLocalMessages',
+              _buildParam({
+                'filter': {
+                  'sourceID': sourceID,
+                  'sessionType': sessionType,
+                  'keywordList': keywordList,
+                  'keywordListMatchType': keywordListMatchType,
+                  'senderUserIDList': senderUserIDList,
+                  'messageTypeList': messageTypeList,
+                  'searchTimePosition': searchTimePosition,
+                  'searchTimePeriod': searchTimePeriod,
+                  'pageIndex': pageIndex,
+                  'count': count,
+                },
+                'operationID': Utils.checkOperationID(operationID),
+              }))
+          .then((value) =>
+              Utils.toObj(value, (map) => SearchResult.fromJson(map)));
+
+  /// Delete message from local and service
+  /// 删除消息
+  Future<dynamic> deleteMessageFromLocalAndSvr({
+    required Message message,
+    String? operationID,
+  }) =>
       _channel.invokeMethod(
-          'searchLocalMessages',
+          'deleteMessageFromLocalAndSvr',
+          _buildParam(message.toJson()
+            ..addAll({
+              "operationID": Utils.checkOperationID(operationID),
+            })));
+
+  /// Delete all message from local
+  /// 删除所有消息
+  Future<dynamic> deleteAllMsgFromLocal({
+    String? operationID,
+  }) =>
+      _channel.invokeMethod(
+          'deleteAllMsgFromLocal',
           _buildParam({
-            'filter': {
-              'sourceID': sourceID,
-              'sessionType': sessionType,
-              'keywordList': keywordList,
-              'keywordListMatchType': keywordListMatchType,
-              'senderUserIDList': senderUserIDList,
-              'messageTypeList': messageTypeList,
-              'searchTimePosition': searchTimePosition,
-              'searchTimePeriod': searchTimePeriod,
-              'pageIndex': pageIndex,
-              'count': count,
-            },
-            'operationID': Utils.checkOperationID(operationID),
+            "operationID": Utils.checkOperationID(operationID),
+          }));
+
+  /// Delete all message from service
+  /// 删除所有消息
+  Future<dynamic> deleteAllMsgFromLocalAndSvr({
+    String? operationID,
+  }) =>
+      _channel.invokeMethod(
+          'deleteAllMsgFromLocalAndSvr',
+          _buildParam({
+            "operationID": Utils.checkOperationID(operationID),
+          }));
+
+  /// Mark conversation message as read
+  /// 标记消息已读
+  Future markMessageAsReadByConID({
+    required String conversationID,
+    required List<String> messageIDList,
+    String? operationID,
+  }) =>
+      _channel.invokeMethod(
+          'markMessageAsReadByConID',
+          _buildParam({
+            "messageIDList": messageIDList,
+            "conversationID": conversationID,
+            "operationID": Utils.checkOperationID(operationID),
           }));
 
   static Map _buildParam(Map param) {